Initial push. Set up the scaffold of the API
Installed the needed NuGet packages for developing the GraphQL API.
Set up DAO, to access the MongoDB database. Set up BookService, which is in charge of accessing and managing the books in the DB. Also introduced GrpahQL special type to represent the Book model, but in the GraphQL schema.
Problem in Program.Cs, to be fixed later!!!!