Added new resolvers on the BookQueryType
Added two new resolvers on the BookQueryType. One resolver that takes in an argument title and returns book based on its title and one resolver that takes in an argument ISBN and returns a book based on it's ISBN. Fixed a minor error with the Type<> of one of the resolvers and changed it to <ListType<GBook>> instead of just <GBook>.
Added a new service called BooksToReadService, that is to be implemented in the Graph query type later. The service is in charge of taking care of the books that are in the collection "BooksToRead".