Added a mutation type and changed the way queries are defined
Changed the way the queries are defined in the query type, for a more readable and understandable code practice.
Designated a directory for specifically mutations. Added a new service in the BookService, that's in charge of handling the mutations of books (creation of books in the Mongo DB).
Added a Mutation type with just one method.